requirements.txt
google-genai
python-dotenv
chromadb
GitPython
langchain-text-splitters
pydantic
streamlit